Check for RScript.exe
Also another check added to install all the packages which are not
present in the system by default.

So first time when this script runs, the user will be prompted to select
the RScript.exe.

Additionally it tries to check whether the required packages exist, if
it doesnt, it tries to install it.

Use returnRexe function to get the instance of the File object pointing
to the RScript.exe in the backbone